Handle multiple directories, can exit on rm prompt

+++ librarian.rb
@@ -4,38 +4,40 @@
 require 'find'
 require 'fileutils'
 
-library = ARGV[0]
+# currently, it just helps you track down errant extensions in your library directories
 
-extensions = {}
+ARGV.each do |library|
+  extensions = {}
 
-Find.find(library) do |path|
-  unless FileTest.directory? path
-    ext = File.extname(path).downcase[1..-1]
-    if ext.nil?
-      ext = '"nil"'
+  Find.find(library) do |path|
+    unless FileTest.directory? path
+      ext = File.extname(path).downcase[1..-1]
+      if ext.nil?
+        ext = '"nil"'
+      end
+      extensions[ext] ||= []
+      extensions[ext] << path
     end
-    extensions[ext] ||= []
-    extensions[ext] << path
   end
-end
 
-puts "Summary:"
-extensions.each { |ext, paths| puts "#{ext}\t#{paths.size}" }
+  puts "Summary:"
+  extensions.each { |ext, paths| puts "#{ext}\t#{paths.size}" }
 
-STDERR.print "Show which extension? "
-while ext = STDIN.gets
-  ext = ext.strip.downcase
-  if extensions[ext]
-    puts extensions[ext]
-    STDERR.print "Delete? [yN] "
-    if STDIN.gets.strip =~ /^y(es)?$/i
-      extensions[ext].each { |path| FileUtils.rm(path) }
-      extensions.delete(ext)
-    end
-  else
-    STDERR.puts "Unknown extension."
-  end
   STDERR.print "Show which extension? "
-end
+  while ext = STDIN.gets
+    ext = ext.strip.downcase
+    if extensions[ext]
+      puts extensions[ext]
+      STDERR.print "Delete? [yN] "
+      if del = STDIN.gets and del.strip =~ /^y(es)?$/i
+        extensions[ext].each { |path| FileUtils.rm(path) }
+        extensions.delete(ext)
+      end
+    else
+      STDERR.puts "Unknown extension."
+    end
+    STDERR.print "Show which extension? "
+  end
 
-puts
+  puts
+end
